icinga2: rename eventstream package to icinga2
