+++ internal/icinga2/api_responses.go
@@ -1,4 +1,4 @@
-package eventstream
+package icinga2
 
 import (
 	"encoding/json"
