+++ internal/icinga2/client_api.go
@@ -1,4 +1,4 @@
-package eventstream
+package icinga2
 
 import (
 	"bufio"
